requirements.txt
alembic==1.11.3
amqp==5.1.1
annotated-types==0.5.0
anyio==3.7.1
astunparse==1.6.3
async-timeout==4.0.3
backports.cached-property==1.0.2
bcrypt==4.0.1
beanie==1.21.0
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.7
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.3
Deprecated==1.2.14
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.103.0
FastAPI-SQLAlchemy==0.2.1
flower==1.2.0
graphql-core==3.2.3
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
humanize==4.6.0
idna==3.4
importlib-metadata==4.13.0
importlib-resources==5.12.0
iniconfig==2.0.0
kombu==5.2.4
lazy-model==0.1.0b0
libcst==1.0.1
Mako==1.2.4
markdown-it-py==2.2.0
MarkupSafe==2.1.3
mdurl==0.1.2
motor==3.3.0
mypy-extensions==1.0.0
packaging==23.1
passlib==1.7.4
pluggy==1.2.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psycopg2==2.9.7
psycopg2-binary==2.9.7
pyasn1==0.5.0
pycparser==2.21
pydantic==2.3.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
Pygments==2.16.1
pymongo==4.5.0
pytest==7.4.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0.1
redis==4.3.4
requests==2.28.1
rich==13.5.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
strawberry-graphql==0.196.2
toml==0.10.2
tomli==2.0.1
tornado==6.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.22.0
vine==5.0.0
wcwidth==0.2.6
wrapt==1.15.0
zipp==3.15.0
